Feature/updated pokemon structs

+++ include/battle_ai_script_commands.h
@@ -12,9 +12,9 @@ void BattleAI_HandleItemUseBeforeAISetup(void);
 void BattleAI_SetupAIData(void);
 u8 BattleAI_ChooseMoveOrAction(void);
 void ClearBankMoveHistory(u8 bank);
-void RecordAbilityBattle(u8 bank, u16 abilityId);
+void RecordAbilityBattle(u32 bank, u32 abilityId);
 void ClearBankAbilityHistory(u8 bank);
-void RecordItemEffectBattle(u8 bank, u8 itemEffect);
+void RecordItemEffectBattle(u32 bank, u32 itemEffect);
 void ClearBankItemEffectHistory(u8 bank);
 u8 BattleAI_ChooseMoveOrAction(void);
